Dependency separation controller and model
